requirements.txt
playwright==1.48.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
mysql-connector-python==9.1.0
pydantic==2.9.2
APScheduler==3.10.4
loguru==0.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.0
openai>=1.0.0
tiktoken>=0.6.0
langchain-community>=0.0.10
